+++ services/web-app/components/catalog/catalog.js
@@ -4,7 +4,8 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { remove, union } from 'lodash'
+import { get, isArray, remove, set, union } from 'lodash'
+import minimatch from 'minimatch'
 import PropTypes from 'prop-types'
 import { useEffect, useState } from 'react'
 import { libraryPropTypes } from 'types'
@@ -15,24 +16,42 @@ import CatalogPagination from '@/components/catalog-pagination'
 import CatalogResults from '@/components/catalog-results'
 import CatalogSearch from '@/components/catalog-search'
 import CatalogSort from '@/components/catalog-sort'
-import { assetSortComparator, librarySortComparator } from '@/utils/schema'
+import {
+  assetSortComparator,
+  collapseAssetGroups,
+  getBaseIdentifier,
+  getCanonicalLibraryId,
+  librarySortComparator
+} from '@/utils/schema'
 import { queryTypes, useQueryState } from '@/utils/use-query-state'
 
 import styles from './catalog.module.scss'
+
+const valuesIntersect = (arr1, arr2) => {
+  if (!isArray(arr1) || !isArray(arr2)) return false
+
+  return arr1.filter((v) => arr2.includes(v)).length
+}
 
 const assetIsInFilter = (asset, filter) => {
   for (const [key, value] of Object.entries(filter)) {
     if (key === 'sponsor') {
       if (!value.includes(asset.params[key])) return false
+    } else if (key === 'tags' && !valuesIntersect(value, asset.content[key])) {
+      return false
     } else {
       if (!value.includes(asset.content[key])) return false
     }
   }
 
+  if (collapseAssetGroups(asset, filter)) {
+    return get(asset, 'library.content.id') === getCanonicalLibraryId(asset)
+  }
+
   return true
 }
 
-function Catalog({ data, filter: defaultFilter = { status: ['stable'] }, type = 'component' }) {
+function Catalog({ collection, data, type, filter: defaultFilter = {}, glob = {} }) {
   const [query, setQuery] = useQueryState('q', {
     defaultValue: ''
   })
@@ -56,7 +75,7 @@ function Catalog({ data, filter: defaultFilter = { status: ['stable'] }, type = 
 
   const [pageSize, setPageSize] = useQueryState('items', {
     ...queryTypes.integer,
-    defaultValue: 12
+    defaultValue: 60
   })
 
   const [filter, setFilter] = useQueryState('filter', {
@@ -68,53 +87,71 @@ function Catalog({ data, filter: defaultFilter = { status: ['stable'] }, type = 
     data.libraries.filter((library) => library.assets.length).sort(librarySortComparator)
   )
 
-  const [initialAssets] = useState(
-    libraries.reduce((allAssets, library) => {
-      return allAssets.concat(
-        library.assets.map((asset) => ({
-          ...asset,
-          library: {
-            params: library.params,
-            content: library.content
-          }
-        }))
-      )
-    }, [])
-  )
-
-  const [filteredAssets, setFilteredAssets] = useState(initialAssets)
-  const [assets, setAssets] = useState(filteredAssets)
-
-  useEffect(() => {
-    setFilteredAssets(
-      initialAssets.filter((asset) => {
+  const [assets] = useState(() => {
+    return libraries
+      .reduce((assetsArray, library) => {
+        // flatten all asset into a single array and save library data per asset
+        return assetsArray.concat(
+          library.assets.map((asset) => ({
+            ...asset,
+            library: {
+              params: library.params,
+              content: library.content
+            }
+          }))
+        )
+      }, [])
+      .filter((asset) => {
         // don't show private assets or assets of the wrong type
-        if (asset.content.private || asset.content.type !== type) return false
+        if (asset.content.private || (type && asset.content.type !== type)) return false
 
-        // don't show if the asset doesn't have one of the valid values
-        return assetIsInFilter(asset, filter)
-      })
-    )
-    // avoid deep object equality comparison in the effect by using JSON.stringify
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [initialAssets, JSON.stringify(filter), type])
-
-  useEffect(() => {
-    setAssets(
-      filteredAssets.sort(assetSortComparator(sort)).filter((asset) => {
-        const { description = '', name = '' } = asset.content
-
-        if (search) {
-          return (
-            (name && name.toLowerCase().includes(search.toLowerCase())) ||
-            (description && description.toLowerCase().includes(search.toLowerCase()))
-          )
+        // don't show libraries or assets if they don't match the glob
+        if (glob.data && glob.pattern) {
+          return minimatch(get(asset, glob.data), glob.pattern)
         }
 
         return true
       })
+  })
+
+  const [filteredAssets, setFilteredAssets] = useState(assets)
+
+  const [assetCounts] = useState(() => {
+    const totals = {}
+
+    assets.forEach((asset) => {
+      const baseIdentifier = getBaseIdentifier(asset)
+
+      if (baseIdentifier) {
+        set(totals, baseIdentifier, get(totals, baseIdentifier, 0) + 1)
+      }
+    })
+
+    return totals
+  })
+
+  useEffect(() => {
+    setFilteredAssets(
+      assets
+        .sort(assetSortComparator(sort))
+        .filter((asset) => assetIsInFilter(asset, filter))
+        .filter((asset) => {
+          const { description = '', name = '' } = asset.content
+
+          if (search) {
+            return (
+              (name && name.toLowerCase().includes(search.toLowerCase())) ||
+              (description && description.toLowerCase().includes(search.toLowerCase()))
+            )
+          }
+
+          return true
+        })
     )
-  }, [filteredAssets, sort, search])
+
+    // avoid deep object equality comparison in the effect by using JSON.stringify
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [assets, JSON.stringify(filter), sort, search])
 
   const handleFilter = (item, key, action = 'add') => {
     let updatedFilter = Object.assign({}, filter)
@@ -149,16 +186,28 @@ function Catalog({ data, filter: defaultFilter = { status: ['stable'] }, type = 
       <CatalogSearch
         className={styles.search}
         filter={filter}
+        initialFilter={{ collection, type }}
         search={search}
         onSearch={handleSearch}
         onFilter={handleFilter}
       />
-      <CatalogFilters filter={filter} onFilter={handleFilter} />
-      <CatalogResults assets={assets} />
+      <CatalogFilters
+        filter={filter}
+        initialFilter={{ collection, type }}
+        onFilter={handleFilter}
+      />
+      <CatalogResults assets={filteredAssets} />
       <CatalogSort onSort={setSort} onView={setView} sort={sort} view={view} />
-      <CatalogList assets={assets} isGrid={view === 'grid'} page={page} pageSize={pageSize} />
+      <CatalogList
+        assetCounts={assetCounts}
+        assets={filteredAssets}
+        isGrid={view === 'grid'}
+        filter={filter}
+        page={page}
+        pageSize={pageSize}
+      />
       <CatalogPagination
-        assets={assets}
+        assets={filteredAssets}
         page={page}
         pageSize={pageSize}
         setPage={setPage}
@@ -169,10 +218,12 @@ function Catalog({ data, filter: defaultFilter = { status: ['stable'] }, type = 
 }
 
 Catalog.propTypes = {
+  collection: PropTypes.string,
   data: PropTypes.shape({
     libraries: PropTypes.arrayOf(libraryPropTypes)
   }),
   filter: PropTypes.object,
+  glob: PropTypes.object,
   type: PropTypes.string
 }