+++ services/web-app/config/constants.js
@@ -8,3 +8,10 @@ import { getRunMode, RunMode } from '@carbon-platform/api/runtime'
 
 export const CACHE_PATH = getRunMode() === RunMode.Prod ? '.carbon' : 'public/data'
 export const IMAGES_CACHE_PATH = '.carbon'
+
+export const ALLOWED_REFERERS = [
+  { protocol: 'https:', domain: '.carbondesignsystem.com' },
+  { protocol: 'https:', domain: 'web-app.j73b4w218e4.us-south.codeengine.appdomain.cloud' },
+  { protocol: 'http:', domain: 'localhost' },
+  { protocol: 'https:', domain: 'localhost' }
+]