+++ services/web-app/lib/file-cache.js
@@ -63,8 +63,7 @@ const slugifyRequest = (host, route, options = {}) => {
  * @returns {Object} GitHub API response data
  */
 const _getResponse = async (host, route, options) => {
-  const responseKey = slugifyRequest(host, route, options)
-  console.log('CACHE MISS', responseKey)
+  // console.log('CACHE MISS', responseKey)
 
   const octokitRef = host === 'github.ibm.com' ? octokitIbm : octokit
 
@@ -86,8 +85,7 @@ const _getResponse = async (host, route, options) => {
  */
 export const getResponse = (host, route, options) => {
   const responseKey = slugifyRequest(host, route, options)
-
-  console.log('CACHE HIT', responseKey)
+  // console.log('CACHE HIT', responseKey)
 
   return diskCache.wrap(responseKey, () => {
     return _getResponse(host, route, options)