+++ services/web-app/next.config.js
@@ -10,6 +10,11 @@ const path = require('path')
 const libraries = require('./data/libraries')
 
 module.exports = {
+  experimental: {
+    outputStandalone: true,
+    // this includes files from the monorepo base two directories up
+    outputFileTracingRoot: path.join(__dirname, '..', '..')
+  },
   i18n: {
     locales: ['en-US'],
     defaultLocale: 'en-US'
@@ -28,6 +33,7 @@ module.exports = {
       @use '~@carbon/react/scss/type' as *;
       @use '~@carbon/react/scss/zone';
       @use './styles/mixins' as *;
+      @use '~carbon-components/scss/globals/scss/_layout' as layout;
     `
   },
   swcMinify: true,
@@ -59,6 +65,12 @@ module.exports = {
   },
   async redirects() {
     return [
+      // temporarily redirect home page for the first release
+      {
+        source: '/',
+        destination: '/assets',
+        permanent: false
+      },
       {
         source: '/assets/:host/:org/:repo/:library',
         destination: '/assets/:host/:org/:repo/:library/latest',