+++ services/web-app/pages/api/logout.js
@@ -7,16 +7,21 @@
 import requireSession from '../../middleware/requireSession'
 
 const logout = requireSession().get((req, res) => {
-  req.logout()
   let nextRoute = '/'
 
   if (req.session.next) {
     nextRoute = req.session.next
     delete req.session.next
   }
-
-  res.redirect(nextRoute)
-  res.end('')
+  req.session.destroy(function () {
+    // clear session cookie
+    res.setHeader(
+      'Set-Cookie',
+      'connect.sid=deleted; path=/; expires=Thu, 01 Jan 1970 00:00:00 GMT'
+    )
+    res.redirect(nextRoute)
+    res.end('')
+  })
 })
 
 export default logout