+++ services/web-app/pages/assets/libraries.js
@@ -4,7 +4,7 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { Download_02 as Download } from '@carbon/pictograms-react'
+import { FileBackup } from '@carbon/pictograms-react'
 import Link from 'next/link'
 import { NextSeo } from 'next-seo'
 import PropTypes from 'prop-types'
@@ -36,7 +36,7 @@ const Libraries = ({ librariesData }) => {
   return (
     <>
       <NextSeo {...seo} />
-      <PageHeader title={seo.title} pictogram={Download} />
+      <PageHeader title={seo.title} pictogram={FileBackup} />
       <ul className={styles.content}>
         {libraries.map((library, i) => (
           <li key={i}>