+++ services/web-app/pages/assets/patterns.js
@@ -4,7 +4,6 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { Dashboard } from '@carbon/pictograms-react'
 import { NextSeo } from 'next-seo'
 import PropTypes from 'prop-types'
 import { useContext, useEffect } from 'react'
@@ -13,6 +12,7 @@ import { libraryPropTypes } from 'types'
 import Catalog from '@/components/catalog'
 import PageHeader from '@/components/page-header'
 import { assetsNavData } from '@/data/nav-data'
+import { type } from '@/data/type'
 import { LayoutContext } from '@/layouts/layout'
 import { getAllLibraries } from '@/lib/github'
 
@@ -30,7 +30,7 @@ const Patterns = ({ librariesData }) => {
   return (
     <>
       <NextSeo {...seo} />
-      <PageHeader title={seo.title} pictogram={Dashboard} />
+      <PageHeader title={seo.title} pictogram={type.pattern.icon} />
       <Catalog data={librariesData} type="pattern" />
     </>
   )